Refactor: add/components common제거, 아톰구조 생성
