+++ src/app/(game)/reactionSpeedGame/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import StartCountdownWrapper from "@/components/common/GameStartWrapper/GameStartWrapper";
+import StartCountdownWrapper from "@/components/molecules/GameStartWrapper";
 import ReactionGameContent from "@/app/(game)/reactionSpeedGame/_components/GameContent";
 
 export default function ReactionSpeadGame() {
