+++ src/app/(game)/wordMathGame/_components/Inputs.tsx
@@ -1,4 +1,4 @@
-import Input from "@/components/common/Input/Input";
+import Input from "@/components/atoms/Input";
 import { ChangeEvent } from "react";
 
 interface InputProps {
